requirements.txt
setuptools==75.6.0
SpeechRecognition==3.10.4
PyAudio==0.2.14
pyttsx3==2.99
sounddevice==0.5.1
vosk==0.3.45
websockets==12.0
psutil==6.1.1
python-dotenv==1.2.1
